[builds] Add some global templates and setup scripts
There's a lot of repetition here, so this sets up a template for builds
and adds some scripts to use to setup new builds and update shared
scripts.

Most of the templates are seeded into new builds then committed, while
the scripts are intended to be largely standard so are excluded from the
default `.gitignore` for each subdirectory.