[templates] Add a configure script
This adds a simple configure script into the templates and the root of
each current build directory. It's intended to be locally editable, as
there may be odd things about a build that need setup, but at the same
time to offer a standard interface that's predictable - it follows the
familiar `./configure; make` pattern.