[images] [base] Add debhelper to the base image
The `debhelper` package clearly needs to be in the base image.

This will use the version from backports on older operating systems as
many package builds require more recent versions - currently this means
Debain Stretch and Ubuntu Xenial and Bionic.

This commit also removes individual installs of this package from some
of the package builds where there was repetition creeping in.