[packages] [libgeo-shapefile-perl] Add libgeo-shapefile-perl build
This adds a build for libgeo-shapefile-perl, pinned at the version
currently in-use (2.52). This package was unmaintained for a long time
(12 years) but picked up a new maintainer in 2019 and has since had a
few updated releases.

I don't think it's used widely enough at the moment to justify testing
an updated version, but this might be something to consider in future.

See: https://metacpan.org/pod/Geo::ShapeFile