[packages] [php-xapian] Update to support PHP 8 based distros
This adds support for Debian Bookworm and Trixie and, for now, removes
support for any other distribution.

Since Xapian 1.4.22 support for PHP < 8 has been dropped but this is not
yet reflected in the source package with the Debian package build config
so for now we just use the bindings source along with fpm.

See: https://xapian.org/docs/bindings/php/