[python] Remove pinned python version and refactor venv handling
This removes the pinned python version and changes the location of the
venv. As a consequence, it excludes the `configure` script from git as
this is now going to be templated, so to permit local changes to the
build configuration for a particular build this now looks for and runs
`bin/local.sh` if it's executable.

Fixes #5