[actions] [python] Add a GitHub action to test script/update
This adds a simple action to set up different Python versions and check
that `script/update` runs. For now, we just expect this to exit 0 but
could add more detailed checks if needed.

This is intended just to give us some confidence that the script will
run on the versions of Python shipped with currently supported Debian
distros.

Fixes #5